+++ containers/daap-query-data/tests/unit/conftest.py
@@ -1,0 +1,44 @@
+import os
+import sys
+from dataclasses import dataclass
+from os.path import dirname, join
+from unittest.mock import Mock
+
+import pytest
+
+sys.path.append(join(dirname(__file__), "../", "../", "src", "var", "task"))
+sys.path.append(
+    join(
+        dirname(__file__), "../", "../", "../", "daap-python-base", "src", "var", "task"
+    )
+)
+
+os.environ["AWS_ACCESS_KEY_ID"] = "testing"
+os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
+os.environ["AWS_SECURITY_TOKEN"] = "testing"
+os.environ["AWS_SESSION_TOKEN"] = "testing"
+os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
+os.environ["BUCKET_NAME"] = "bucket"
+
+
+@dataclass
+class FakeContext:
+    function_name: str
+
+
+@pytest.fixture
+def fake_context():
+    """
+    Emulate the context object passed
+    """
+    return FakeContext(function_name="some-function")
+
+
+@pytest.fixture
+def athena_client():
+    """
+    Create a mock athena service
+    """
+    # Mock the Athena client and its methods
+    athena_client = Mock()
+    return athena_client