+++ terraform/github/application-migration/outputs.tf
@@ -1,4 +1,0 @@
-output "team_list" {
-  value = transpose(local.team_repo_map)
-}
-